+++ framework/Furion/JsonSerialization/Providers/SystemTextJsonSerializerProvider.cs
@@ -1,9 +1,5 @@
 ﻿using Furion.DependencyInjection;
-using System;
-using System.IO;
 using System.Text.Json;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace Furion.JsonSerialization
 {
@@ -25,20 +21,6 @@ namespace Furion.JsonSerialization
         }
 
         /// <summary>
-        /// 序列化流
-        /// </summary>
-        /// <param name="utf8Json"></param>
-        /// <param name="value"></param>
-        /// <param name="inputType"></param>
-        /// <param name="jsonSerializerOptions"></param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        public Task SerializeAsync(Stream utf8Json, object value, Type inputType, object jsonSerializerOptions = null, CancellationToken cancellationToken = default)
-        {
-            return JsonSerializer.SerializeAsync(utf8Json, value, inputType, jsonSerializerOptions as JsonSerializerOptions, cancellationToken);
-        }
-
-        /// <summary>
         /// 反序列化字符串
         /// </summary>
         /// <typeparam name="T"></typeparam>
@@ -48,19 +30,6 @@ namespace Furion.JsonSerialization
         public T Deserialize<T>(string json, object jsonSerializerOptions = null)
         {
             return JsonSerializer.Deserialize<T>(json, jsonSerializerOptions as JsonSerializerOptions);
-        }
-
-        /// <summary>
-        /// 反序列化流
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="utf8Json"></param>
-        /// <param name="jsonSerializerOptions"></param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        public ValueTask<T> DeserializeAsync<T>(Stream utf8Json, object jsonSerializerOptions = null, CancellationToken cancellationToken = default)
-        {
-            return JsonSerializer.DeserializeAsync<T>(utf8Json, jsonSerializerOptions as JsonSerializerOptions, cancellationToken);
         }
     }
 }